Move User from controller-api to jdisc-security-filters
