+++ jdisc-security-filters/src/main/java/com/yahoo/jdisc/http/filter/security/misc/User.java
@@ -1,5 +1,5 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.hosted.controller.api.integration.user;
+package com.yahoo.jdisc.http.filter.security.misc;
 
 import java.time.LocalDate;
 import java.util.Objects;
